remove AbstractDatabase
